// Web/src/app/configuration/project-config/project/project-store/project-action-type.enum.ts
export enum ProjectActionType {
    FetchTicketSystemTypes = 'FETCH_TICKEYSYSTEM_TYPES',
    ShowDivisionControl = 'SHOW_DIVISION_CONTROL',
    ShowCodeAnalysisDiv = 'SHOW_CODE_ANALYSIS_DIV',
    CurrentTicketSystemType = 'CURRENT_TICKETSYSTEM_TYPE',
    EnterpriseAccountState = 'ENTERPRISE_ACCOUNT_STATE',
    FetchProjectInfo = 'FETCH_PROJECT_INFO',
    FetchCodeAnalysisInfo = 'FETCH_CODE_ANALYSIS_INFO',
    FetchAppAnalysisInfo = 'FETCH_APP_ANALYSIS_INFO',
    ShowTab = 'SHOW_TAB',
    FetchStaticMappingInfo = 'FETCH_STATIC_MAPPING_INFO',
    FetchTicketSystemInfo = 'FETCH_TICKET_SYSTEM_INFO',
    ShowUserMapTab = 'SHOW_USER_MAP_TAB',
    FetchRoleOptions = 'FETCH_ROLES'
}